Consolidate Sv Computing into a single module (#153)
