+++ oceanstream/echodata/__init__.py
@@ -1,6 +1,5 @@
-from .compute_sv import compute_sv_with_encode_mode
 from .processed_data_io import read_processed, write_processed
 from .read_file import read_file
-from .sv_computation import compute_sv
+from .sv_computation import compute_sv, compute_sv_with_encode_mode
 from .sv_dataset_extension import enrich_sv_dataset
 from .sv_interpolation import interpolate_sv, regrid_dataset
